Polygon function
I worked on the function "polygon" to command tello to take a specific flight path based off of initial parameters. It appears to pass some, but not all commands to tello, suggesting that I might need to add a delay between commands.

Future work on this notebook will include fixing this function, and measuring drift of the drone with the mission pads.