Added drift test
I have added test of how far the drone drifts from a hovering point, testing how long it takes to return to that origin after drifting.

This method was selected because as of this time I am not sure how to obtain coordinate information from the tello drone, only how to send that info to the drone.

In the future I will complile data with this test to determine how much the drone drifts without external wind. Or if I can find a method to get coordinate information from the drone I might implement that instead.